TelegramBotResponder/requirements.txt
paho_mqtt==1.6.1
Requests==2.31.0
telepot==12.7
